refactor: deduplicate shared helpers into hook-settings and templates modules
Consolidate isOpenWolfHook, replaceOpenWolfHooks, and findTemplatesDir which
were duplicated across init.ts and update.ts — each now lives in one canonical
module. Fixes path.basename("") bug in post-write.ts that produced empty
filenames in bug-log summaries, adds worktree guard to update command, and
removes redundant process.exit mock from session-start tests.
