+++ src/cli/hook-settings.ts
@@ -39,6 +39,48 @@ export const HOOK_SETTINGS = {
   ],
 };
 
+/**
+ * Returns true if a hook entry was registered by OpenWolf
+ * (i.e., its command references .wolf/hooks/).
+ */
+export function isOpenWolfHook(hook: unknown): boolean {
+  if (typeof hook !== "object" || hook === null) return false;
+  const h = hook as Record<string, unknown>;
+  if (typeof h.command === "string" && h.command.includes(".wolf/hooks/")) return true;
+  return false;
+}
+
+/**
+ * Replace OpenWolf hooks in an existing settings object while preserving
+ * any user-added hooks that are NOT OpenWolf hooks.
+ */
+export function replaceOpenWolfHooks(
+  existing: Record<string, unknown>,
+  newHooks: typeof HOOK_SETTINGS
+): Record<string, unknown> {
+  const merged: Record<string, unknown> = { ...existing };
+  const existingHooks = (typeof existing.hooks === "object" && existing.hooks !== null)
+    ? { ...(existing.hooks as Record<string, unknown>) }
+    : {};
+
+  for (const event of Object.keys(newHooks) as Array<keyof typeof HOOK_SETTINGS>) {
+    const existing_entries = Array.isArray(existingHooks[event])
+      ? (existingHooks[event] as unknown[])
+      : [];
+    // Keep non-OpenWolf entries the user may have added
+    const userEntries = existing_entries.filter((entry) => {
+      if (typeof entry !== "object" || entry === null) return true;
+      const e = entry as Record<string, unknown>;
+      const hooks = Array.isArray(e.hooks) ? e.hooks : [];
+      return !hooks.some(isOpenWolfHook);
+    });
+    existingHooks[event] = [...newHooks[event], ...userEntries];
+  }
+
+  merged.hooks = existingHooks;
+  return merged;
+}
+
 // Hook script basenames that get copied from dist/hooks/ into .wolf/hooks/.
 // Single source of truth — duplication across init/update/status caused the
 // worktree-helper.js bug surfaced in PR #25 review.
